Add track component
