+++ plugins/data-management/src/AddTrackWidget/components/TrackTypeSelector.tsx
@@ -1,7 +1,6 @@
 import { getEnv } from '@jbrowse/core/util'
 import { MenuItem, TextField } from '@mui/material'
 import { observer } from 'mobx-react'
-import { makeStyles } from 'tss-react/mui'
 
 import type { AddTrackModel } from '../model'
