+++ plugins/data-management/src/AddTrackWidget/components/doSubmit.ts
@@ -1,4 +1,3 @@
-import { getConf } from '@jbrowse/core/configuration'
 import {
   getSession,
   isElectron,
